requirements.txt
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.9.0
nest-asyncio>=1.5.0
pyyaml>=6.0
pytest>=7.4.0
matplotlib>=3.7.0
